requirements.txt
aiohttp >= 3.7.4.post0
colorama >= 0.4.4
discord.py >= 1.7.3
discord >= 1.7.3
ruamel.yaml >= 0.17.16
tqdm >= 4.62.3
